@@ -17,7 +17,7 @@
     "prerelease": "npm run build",
     "release": "sh ./scripts/release.sh",
     "serve:docs": "compodoc --serve --open --output=\"docs/\"",
-    "start": "ng serve --aot -o --port 5005 --host 0.0.0.0",
+    "start": "ng serve --aot --open --port 5005 --host 0.0.0.0",
     "test": "ng test",
     "test:single-run": "npm run lint && ng test --watch false"
   },
@@ -40,50 +40,51 @@
     "bootstrap4-paginator"
   ],
   "peerDependencies": {
-    "@angular/common": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "bootstrap": "^4.1.1",
-    "rxjs": "^5.5.6"
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "bootstrap": "^4.1.3",
+    "rxjs": "~6.2.0"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cli": "1.7.3",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@compodoc/compodoc": "^1.1.3",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "bootstrap": "^4.1.1",
-    "codelyzer": "^4.0.1",
-    "commitizen": "^2.9.6",
-    "core-js": "^2.4.1",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/cli": "~6.2.2",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@compodoc/compodoc": "^1.1.5",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "bootstrap": "^4.1.3",
+    "codelyzer": "~4.3.0",
+    "commitizen": "^2.10.1",
+    "core-js": "^2.5.4",
     "cz-conventional-changelog": "^2.1.0",
     "font-awesome": "^4.7.0",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^2.4.5",
-    "protractor": "~5.1.2",
-    "rxjs": "^5.5.6",
-    "ts-node": "~4.1.0",
-    "tsickle": "^0.27.2",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "zone.js": "^0.8.19"
+    "ng-packagr": "^4.1.1",
+    "protractor": "~5.4.0",
+    "rxjs": "~6.2.0",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.32.1",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2",
+    "zone.js": "~0.8.26"
   },
   "config": {
     "commitizen": {
